Refactored renderer.js
