+++ utils.js
@@ -1,0 +1,4 @@
+// utils.js
+export async function resolveContent(content) {
+    return content instanceof Promise ? await content : content;
+  }
